fix(web-app): localhost fetch requests
TLS termination on the deployed prod app happens at a layer outside of the actual running
application.

Internal requests should be made as normal http requests since they do not leave the confines of the
running container.

The exception to this is if the service is explicitly running securely using secure-dev-server.js